Supports db/image config option for PostgreSQL
Currently for PostgreSQL database containers only, supports an optional "image" setting under the "db" settings in a project's .up.json.

The "image" setting can specify a custom Docker image and tag (eg "studionone/postgres:9.5").

This setting is optional. If not specified, the default of "postgres:9.5" (the official PostgreSQL 9.5 image) used, as before.