v1.0.6 release: Fixes Docker version number issues
This changes the logic involved in checking for a minimum Docker version. Previously, we had a maximum version as well; with the pending release of v17.3 this needed to be updated to be more robust.

Tests for the change, including the upcoming version numbers have been added to `test/util_test.nim`, and pass.

Version has been bumped to v1.0.6